Row buttons for the actions a row already owns (#218)
* feat(ui): drop the card override dot, prune no-op overrides

Home's card title and menu no longer draw the profile-bound dot; the
collections list drops its current-collection mark dot (the cursor already
opens on that row).

A profile row edited back to the global value clears its override instead of
pinning an equal value.

* feat(ui): row buttons for the actions a row already owns

The host menu's power row carries the ⋯ onto its power settings and Connect
the pencil onto the address behind it; "Connect with…" carries the ⋯ onto the
profile it defaults to. The settings page's calibrate row carries the bin that
throws a calibration away, and only once there is one. Their rows are gone.

Reached the way a collection row's buttons are: one marks table per screen
behind `row_buttons`, Left/Right to step onto them, Confirm keyed by icon. The
pointer paths share one `kit_row_button_at`, and `draws_kit_rows` replaces the
two hand-copied list-membership guards the settings page needed.

Type also rides one scale (`BASE_SCALE`) rather than per-screen sizes.

+++ src/app/draw/settings.rs
@@ -143,7 +143,7 @@ pub(crate) fn draw(
                 c,
                 p.label(),
                 f64::from(r.left + 44.0 * k),
-                f64::from(r.center_y() + ENTRY_SIZE as f32 * k * 0.35),
+                f64::from(r.center_y()) + ENTRY_SIZE * f64::from(k) * 0.35,
                 if open { W::SemiBold } else { W::Medium },
                 ENTRY_SIZE * f64::from(k),
                 color,
